refactor: 提取shiftGridX/Y消除rollover四段重复

+++ src/guga_perception/terrainanalysis/terrain_analysis/include/terrain_analysis/core/context.hpp
@@ -21,11 +21,11 @@ struct TerrainConfig {
   static constexpr int PLANAR_VOXEL_HALF_WIDTH = (PLANAR_VOXEL_WIDTH - 1) / 2;
   static constexpr int PLANAR_VOXEL_NUM = PLANAR_VOXEL_WIDTH * PLANAR_VOXEL_WIDTH;
 
-  static constexpr int terrain_voxel_index(int row, int col) {
-    return TERRAIN_VOXEL_WIDTH * row + col;
+  static constexpr size_t terrain_voxel_index(int row, int col) {
+    return static_cast<size_t>(TERRAIN_VOXEL_WIDTH * row + col);
   }
-  static constexpr int planar_voxel_index(int row, int col) {
-    return PLANAR_VOXEL_WIDTH * row + col;
+  static constexpr size_t planar_voxel_index(int row, int col) {
+    return static_cast<size_t>(PLANAR_VOXEL_WIDTH * row + col);
   }
 
   float scan_voxel_size = 0.05;
